+++ components/ads-client/integration-tests/tests/http_cache.rs
@@ -6,7 +6,7 @@
 use std::hash::{Hash, Hasher};
 use std::time::Duration;
 
-use ads_client::database::bytesize::ByteSize;
+use ads_client::common::bytesize::ByteSize;
 use ads_client::http_cache::{CacheOutcome, CachePolicy, HttpCache};
 use mockito::mock;
 use viaduct::{Client, ClientSettings, Request};
