+++ components/ads-client/src/http_cache/store.rs
@@ -5,7 +5,7 @@
 use std::{collections::HashMap, sync::Arc, time::Duration};
 
 use crate::{
-    database::clock::{CacheClock, Clock},
+    common::clock::{CacheClock, Clock},
     http_cache::{request_hash::RequestHash, ByteSize},
 };
 use parking_lot::Mutex;
@@ -46,7 +46,7 @@ impl HttpCacheStore {
 
     #[cfg(test)]
     pub fn new_with_test_clock(conn: Connection) -> Self {
-        use crate::database::clock::TestClock;
+        use crate::common::clock::TestClock;
 
         Self {
             conn: Mutex::new(conn),
