+++ components/ads-client/src/lib.rs
@@ -16,7 +16,7 @@ use http_cache::CachePolicy;
 use mars::ad_request::{AdPlacementRequest, AdRequestFlags};
 pub mod ads_store;
 mod client;
-pub mod database;
+pub mod common;
 mod ffi;
 pub mod http_cache;
 mod mars;
