+++ components/ads-client/src/mars.rs
@@ -268,7 +268,7 @@ mod tests {
 
         let cache = HttpCache::builder("test_fetch_ads_cache_hit_skips_network.db")
             .default_ttl(std::time::Duration::from_secs(300))
-            .max_size(crate::database::bytesize::ByteSize::mib(1))
+            .max_size(crate::common::bytesize::ByteSize::mib(1))
             .build()
             .unwrap();
         let client = make_test_client(Some(cache));
@@ -306,7 +306,7 @@ mod tests {
         viaduct_dev::init_backend_dev();
         let cache = HttpCache::builder("test_record_click.db")
             .default_ttl(std::time::Duration::from_secs(300))
-            .max_size(crate::database::bytesize::ByteSize::mib(1))
+            .max_size(crate::common::bytesize::ByteSize::mib(1))
             .build()
             .unwrap();
 
@@ -325,7 +325,7 @@ mod tests {
         viaduct_dev::init_backend_dev();
         let cache = HttpCache::builder("test_record_impression.db")
             .default_ttl(std::time::Duration::from_secs(300))
-            .max_size(crate::database::bytesize::ByteSize::mib(1))
+            .max_size(crate::common::bytesize::ByteSize::mib(1))
             .build()
             .unwrap();
